Thermal insulation with and without insulation
