+++ femmt/examples/Three_Winding_Transformer.py
@@ -1,27 +1,27 @@
 import femmt as fmt
 import os
 
-def example_thermal_simulation():
+def example_thermal_simulation(flag_insulation: bool):
     # Thermal simulation:
     # The losses calculated by the magnetics simulation can be used to calculate the heat distribution of the given magnetic component
     # In order to use the thermal simulation, thermal conductivities for each material can be entered as well as a boundary temperature
     # which will be applied on the boundary of the simulation (dirichlet boundary condition).
-    
+
     # The case parameter sets the thermal conductivity for a case which will be set around the core.
     # This could model some case in which the transformer is placed in together with a set potting material.
     thermal_conductivity_dict = {
-            "air": 0.0263,
-            "case": { # epoxy resign
-                "top": 1.54,
-                "top_right": 1.54,
-                "right": 1.54,
-                "bot_right": 1.54,
-                "bot": 1.54
-            },
-            "core": 5, # ferrite
-            "winding": 400, # copper
-            "air_gaps": 180, # aluminiumnitride
-            "insulation": 0.42 # polyethylen
+        "air": 1.54,
+        "case": {  # epoxy resign
+            "top": 1.54,
+            "top_right": 1.54,
+            "right": 1.54,
+            "bot_right": 1.54,
+            "bot": 1.54
+        },
+        "core": 5,  # ferrite
+        "winding": 400,  # copper
+        "air_gaps": 180,  # aluminiumnitride
+        "insulation": 0.42 if flag_insulation else None  # polyethylen
     }
 
     # Here the case size can be determined
@@ -40,7 +40,7 @@ def example_thermal_simulation():
         "value_boundary_bottom_right": 20,
         "value_boundary_bottom": 20
     }
-    
+
     # In order to compare the femmt thermal simulation with a femm heat flow simulation the same boundary temperature should be applied.
     # Currently only one temperature can be applied which will be set on every boundary site.
     femm_boundary_temperature = 20
@@ -62,16 +62,22 @@ def example_thermal_simulation():
     # When the losses file is already created and contains the losses for the current model, it is enough to run geo.create_model in
     # order for the thermal simulation to work (geo.single_simulation is not needed).
     # Obviously when the model is modified and the losses can be out of date and therefore the geo.single_simulation needs to run again.
-    geo.thermal_simulation(thermal_conductivity_dict, boundary_temperatures, boundary_flags, case_gap_top,
-                           case_gap_right, case_gap_bot, True, color_scheme=fmt.colors_ba_jonas, colors_geometry=fmt.colors_geometry_ba_jonas)
+    # example_thermal_simulation(True)  # For simulation with insulation
+    # example_thermal_simulation(False)  # For simulation without insulation
+
+    geo.thermal_simulation(flag_insulation, thermal_conductivity_dict, boundary_temperatures, boundary_flags,
+                           case_gap_top,
+                           case_gap_right, case_gap_bot, True, color_scheme=fmt.colors_ba_jonas,
+                           colors_geometry=fmt.colors_geometry_ba_jonas)
+
 
 example_results_folder = os.path.join(os.path.dirname(__file__), "example_results")
 if not os.path.exists(example_results_folder):
     os.mkdir(example_results_folder)
 
-#component = "inductor"
+component = "inductor"
 #component = "transformer-interleaved"
-component = "transformer"
+#component = "transformer"
 #component = "integrated_transformer"
 # component = "load_from_file"
 
@@ -83,12 +89,19 @@ if component == "inductor":
         os.mkdir(working_directory)
 
     # 1. chose simulation type
-    geo = fmt.MagneticComponent(component_type=fmt.ComponentType.Inductor, working_directory=working_directory, silent=True)
+    geo = fmt.MagneticComponent(component_type=fmt.ComponentType.Inductor, working_directory=working_directory,
+                                silent=False)
+
+    inductor_frequency = 270000
 
     # 2. set core parameters
     core_db = fmt.core_database()["PQ 40/40"]
-    inductor_frequency = 270000
-    core = fmt.Core(core_inner_diameter=core_db["core_inner_diameter"], window_w=core_db["window_w"], window_h=core_db["window_h"],
+    core_dimensions = fmt.dtos.SingleCoreDimensions(core_inner_diameter=core_db["core_inner_diameter"],
+                                                    window_w=core_db["window_w"],
+                                                    window_h=core_db["window_h"])
+
+    core = fmt.Core(core_type=fmt.CoreType.Single,
+                    core_dimensions=core_dimensions,
                     material="N49", temperature=45, frequency=inductor_frequency,
                     # permeability_datasource="manufacturer_datasheet",
                     permeability_datasource=fmt.MaterialDataSource.Measurement,
@@ -97,20 +110,21 @@ if component == "inductor":
                     permittivity_datasource=fmt.MaterialDataSource.Measurement,
                     permittivity_datatype=fmt.MeasurementDataType.ComplexPermittivity,
                     permittivity_measurement_setup="LEA_LK")
-                    # mu_rel=3000, phi_mu_deg=10,
-                    # sigma=0.5)
+    # mu_rel=3000, phi_mu_deg=10,
+    # sigma=0.5)
     geo.set_core(core)
 
     # 3. set air gap parameters
     air_gaps = fmt.AirGaps(fmt.AirGapMethod.Percent, core)
     air_gaps.add_air_gap(fmt.AirGapLegPosition.CenterLeg, 0.0005, 50)
-    #air_gaps.add_air_gap(fmt.AirGapLegPosition.CenterLeg, 0.0005, 90)
+    # air_gaps.add_air_gap(fmt.AirGapLegPosition.CenterLeg, 0.0002, 90)
     geo.set_air_gaps(air_gaps)
 
     # 4. set insulations
-    insulation = fmt.Insulation()
-    insulation.add_core_insulations(0.001, 0.001, 0.004, 0.001)
-    insulation.add_winding_insulations([0.0005], 0.0001)
+    insulation = fmt.Insulation(flag_insulation=False)
+    insulation.add_core_insulations(0.001, 0.001, 0.001, 0.001)
+
+    insulation.add_winding_insulations([[0.0005]])
     geo.set_insulation(insulation)
 
     # 5. create winding window and virtual winding windows (vww)
@@ -121,12 +135,12 @@ if component == "inductor":
     winding = fmt.Conductor(0, fmt.Conductivity.Copper)
     winding.set_solid_round_conductor(conductor_radius=0.0013, conductor_arrangement=fmt.ConductorArrangement.Square)
     winding.parallel = False  # set True to make the windings parallel, currently only for solid conductors
-    #winding.set_litz_round_conductor(conductor_radius=0.0013, number_strands=150, strand_radius=100e-6,
+    # winding.set_litz_round_conductor(conductor_radius=0.0013, number_strands=150, strand_radius=100e-6,
     # fill_factor=None, conductor_arrangement=fmt.ConductorArrangement.Square)
 
     # 7. add conductor to vww and add winding window to MagneticComponent
     vww.set_winding(winding, 9, None)
-    geo.set_winding_window(winding_window)
+    geo.set_winding_windows([winding_window])
 
     # 8. create the model
     geo.create_model(freq=inductor_frequency, pre_visualize_geometry=True, save_png=False)
@@ -146,7 +160,7 @@ if component == "inductor":
     # 9. start simulation
 
     # 7. prepare and start thermal simulation
-    # example_thermal_simulation()
+    example_thermal_simulation(False)
 
 if component == "transformer-interleaved":
     working_directory = os.path.join(example_results_folder, "transformer-interleaved")
